requirements.txt
dj-database-url==0.5.0
django-livereload==1.7
django-livereload-server==0.3.2
gunicorn==20.1.0
pillow==8.2.0
psycopg2-binary==2.9.1
python-decouple==3.4
requests==2.25.1
whitenoise==5.2.0
django-cleanup==5.2.0
